6c amd64: fix immconst() long-vs-int32, unblocking 64-bit CMP immediates
Another instance of this project's recurring long-vs-int32-on-a-64-bit-
host bug: immconst()'s "does this vlong constant fit as a 32-bit signed
immediate" check used `long v`, which is 64 bits on our hosts, so
`v = n->vconst; n->vconst == (vlong)v` never differed and immconst()
always returned true -- even for vlong constants that don't fit in 32
bits. hardconst() (which callers use to decide whether a constant needs
to be materialized into a register first, e.g. in boolgen()'s
comparison codegen) is defined as `!immconst(n)`, so it never fired
either.

The result wasn't just wrong codegen but an outright assembly failure:
comparing a vlong against an out-of-range constant emitted "CMPQ
reg,$1099511627818", and x86-64 has no 64-bit-immediate CMP form (only
32-bit sign-extended), so 6l's doasm() rejected it outright ("notfound").
Found while switching tests/c/mini2's regress7c_macos_amd64.exe off
6cg: its id64(1099511627818LL) == 1099511627818LL comparison (2^40+42)
hits exactly this path; the bug is OS-independent, just previously only
exercised by that macOS-only file. Go-era's own 6c (src/cmd/6c/cgen.c
immconst()) already uses int32 here.

Adds tests/c/variants/vlongimm_amd64.c: a plain-ELF, qemu-runnable
regression test for the same pattern (verified it fails -- with a hard
assembler error, not just wrong output -- against the pre-fix compiler,
passes after).

+++ compilers/6c/cgen.c
@@ -1820,11 +1820,19 @@ layout(Node *f, Node *t, int c, int cv, Node *cn)
 
 /*
  * constant is not vlong or fits as 32-bit signed immediate
+ * claude: was `long v`. On our 64-bit host `long` is 64 bits, so
+ * `v = n->vconst; n->vconst == (vlong)v` never differs and this always
+ * returned true, even for vlong constants that don't fit in 32 bits --
+ * e.g. comparing against 1099511627818LL (2^40+42) generated an
+ * unencodable "CMPQ reg,$1099511627818" (x86-64 has no 64-bit-immediate
+ * CMP form) instead of routing through hardconst()'s register-
+ * materializing path. Go-era's own 6c (src/cmd/6c/cgen.c immconst())
+ * already uses int32 here for exactly this reason -- matching it.
  */
 int
 immconst(Node *n)
 {
-	long v;
+	int32 v;
 
 	if(n->op != OCONST || !typechlpv[n->type->etype])
 		return 0;
